Cache JSON in localStorage
The network request is quite fast, but since the JSON rarely changes, it
probably doesn't need to happen on every page load. I'm looking at
integrating this into MetaCPAN and I think if we can avoid thousands of
unnecessary network requests every day, that might be a little
friendlier to the environment.